requirements.txt
absl-py==0.1.10
astroid==1.6.1
bleach==1.5.0
bottle==0.12.13
h5py==2.7.1
html5lib==0.9999999
isort==4.3.4
Keras==2.1.4
numpy==1.14.0
Pillow==5.0.0
protobuf==3.5.1
scipy==1.0.0
tensorflow==1.5.0
tensorflow-tensorboard==1.5.1
